feat(plc): black-body Serial/TCP+UDP links, mm servo coords, momentary P-bit triggers
Hardware/protocol
- BlackBodyUnitSettings replaces SerialSettings for BlackBodySettings.Units with
  per-unit ConnectionType (Serial/Ip), IpAddress and Port; SrBlackBodyController
  routes each unit to the matching link (UdpSrLink/SerialPortSrLink).
- Servo point coordinates are mm floats end to end; PlcXgtClient scales to and
  from the PLC's 0.1mm words, and PlcStatusSnapshot carries mm.
- P-area trigger bits are momentary: WriteBitAsync writes ON, waits
  PlcSettings.PulseHoldMs (100ms), then writes OFF. JOG keeps hold semantics.
- New alarm triggers: ResetErrorAsync (P525) and BuzzerOffAsync (P250).
- Black-body words use the x100 scale; the simulator ramps them at that scale and
  samples one-touch triggers on a 20ms scan so a 100ms pulse is never missed.

Master UI
- PLC settings: black-body connection card (per-unit type + type-dependent
  Serial/IP rows).
- Manual control gains target temperature/humidity, the temperature ramp, and
  servo/fan speed, all moved out of PLC settings. Fan speed keeps the polled
  value as a read-only current reading beside its target field.
- Main window bottom-left: PLC buzzer OFF and PLC error reset buttons with a
  result line; ko/en strings added.
- Status monitor now consumes the shared PlcStatusService snapshot instead of
  polling ReadStatusAsync itself. Its own 1s poll (~70 tag round-trips) had no
  re-entrancy guard and starved the shared poller on PlcXgtClient's single IO
  semaphore, so the view sat at "폴링 대기" and never showed a value.
- Servo positions render with one decimal and a mm unit.

Tests: 198 pass, including TCP/UDP link round-trips, P-bit pulse and JOG-hold
behavior, coordinate scaling against the FEnet simulation service, and
black-body settings JSON round-trip.

+++ HeatingCameraSystem.Core/Models/PlcModels.cs
@@ -61,8 +61,9 @@ namespace HeatingCameraSystem.Core.Models
         public float BlackBody2Sv { get; set; }
 
         // ── 서보/모션 ──
-        public int ServoXPosition { get; set; }
-        public int ServoYPosition { get; set; }
+        // PLC 워드는 0.1mm 단위 정수(x10 스케일). 여기서는 스케일 해제된 mm 값.
+        public float ServoXPosition { get; set; }
+        public float ServoYPosition { get; set; }
         public bool ServoXBusy { get; set; }
         public bool ServoYBusy { get; set; }
         public bool ServoXHomeComplete { get; set; }
